go/staking/tests/debug: Move into go/staking/tests package
