+++ .changelog/3774.internal.md
@@ -1,0 +1,1 @@
+go/staking/tests/debug: Move into go/staking/tests package
